Inject attribute factory into resolve metadata
This allows non-gradle metadata implementations (Maven and Ivy) to
derive variants with attributes for modules under certain conditions.
This will be used to derive variants in case a module is used as
a Java Library.
To make such decisions, the consumer attributes are now passed to the
variant extraction method getVariantsForGraphTraversal().